feat: implement user preference onboarding and count-based POI recommendation
- Add ReplaceUserPreferences logic to reset and insert user preferences
- Implement UpdatePreferences endpoint for authenticated users
- Add count-based scoring for POI recommendation
- Sort POIs by number of matching user preferences
- Improve geocoding service by adding city parameter for better accuracy
- Optimize preference matching using HashSet
- Load trip with segments and itineraries
- Implement temperature, wind, and precipitation scoring
- Add adaptive weather risk calculation
- Prepare itinerary risk update logic

+++ Domain/Entities/ItineraryDetail.cs
@@ -20,7 +20,7 @@ namespace Domain.Entities
 
         public Guid? PoiId { get; set; }
 
-        public float WeatherRiskScore { get; set; }
+        public double WeatherRiskScore { get; set; }
 
         public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
